fix: minor bug fixes and refactors
