+++ src/components/marginals/Footer/Footer.jsx
@@ -8,14 +8,14 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   MainContainer,
   SubContainer,
-  Container,
+  Container as SomeContainer,
   NavContainer,
   BottomContainer,
   SocialContainer,
   NavLink,
 } from './styles';
 import { Para2, Para1 } from '../..';
-import { SectionContainer } from '../../shared';
+import Container from '../../shared/Container';
 
 // assets
 import { footer, nav } from '../../../../config/content';
@@ -29,14 +29,14 @@ const Footer = () => {
 
   return (
     <MainContainer>
-      <SectionContainer>
+      <Container>
         <SubContainer>
-          <Container>
+          <SomeContainer>
             <Link to='/'>
               <img className='logo' src={logo.src} alt={logo.alt} />
             </Link>
             <Para2 className='sub-heading'>{subHeading}</Para2>
-          </Container>
+          </SomeContainer>
 
           <NavContainer>
             <div className='column'>
@@ -55,12 +55,12 @@ const Footer = () => {
             </div>
           </NavContainer>
 
-          <Container>
+          <SomeContainer>
             <Para1 className='contact-heading'>{contact.heading}</Para1>
             <Para2 className='contact-email'>{contact.body}</Para2>
             <Para2 className='privacy'>{contact.privacy}</Para2>
             <Para2 className='terms'>{contact.terms}</Para2>
-          </Container>
+          </SomeContainer>
         </SubContainer>
 
         <BottomContainer>
@@ -74,7 +74,7 @@ const Footer = () => {
 
           <h5 className='copyright'>{copyright}</h5>
         </BottomContainer>
-      </SectionContainer>
+      </Container>
     </MainContainer>
   );
 };
