+++ src/components/shared/PageContainer.jsx
@@ -1,16 +1,14 @@
+// Libraries
 import styled from 'styled-components';
 
 export default styled.div`
+  width: 100%;
   max-width: 100vw;
   height: auto;
   min-height: 100vh;
+  margin: 0;
+  padding: 0;
   padding-top: 80px;
-
   overflow-x: hidden;
   overflow-y: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 80px 5% 100px;
 `;
